refactor(trends): extract corpus freshness service (#1853)
